requirements.txt
flask
flask_SQLAlchemy
flask_cors
requests
gunicorn
python-dotenv
mysqlclient
bcrypt
lyricsgenius